Dev/denum (#85)
* Phase 1: DEnum runtime metadata + DEnumProperty (#79)

* Add DEnum runtime metadata and DEnumProperty for enum class reflection

Phase 1: introduce DENUM() macro, DEnum metadata with ReflectionRegistry
registration, EPropertyType::Enum, and DEnumProperty<T> with integer JSON
serialization. Includes manual test enum registration and engine tests.


* Fix DEnumProperty compile errors: move to separate header

DEnumProperty was inlined at the bottom of DProperty.h with includes after
the class body, causing incomplete-type errors (ToString before DEnum.h)
and a circular include with AssetArchive.h. Move the template to
DEnumProperty.h which includes dependencies in safe order.


* Fix DeltaEngineTests link errors for DEnumProperty

Export RegisterDEnum from the DLL and add DEnumPropertyBase with an
exported constructor in DEnum.cpp so test code can instantiate
DEnumProperty without resolving the non-exported DProperty ctor.


* Fix MSVC C2487: remove duplicate DLL export on DEnumPropertyBase

Export the constructor and GetEnumSchema individually instead of marking
the whole class DELTAENGINE_API, matching DDelegatePropertyBase pattern.


---------


* Add DENUM() codegen to DeltaHeaderTool (Phase 2) (#80)

Parse scoped enum class declarations annotated with DENUM(), emit DEnum
registration and DEnumProperty<T> bindings, and wire EReflectionTestEnum
into ReflectionTestObject with generated registration replacing manual
test setup in DEnumPropertyTests.


* Add enum property editor UI, undo JSON, and tests (#81)

Wire EPropertyType::Enum through Details combo widget, PropertyValueIO
integer round-trip, and undo eligibility. Adds EnumField widget,
EnumPropertyWire helpers, and editor test coverage for ReflectionTestObject.


* Phase 4: MCP enum schema, vector-of-enum, switch hygiene (#82)

Expose enum metadata (type, enum_name, values) in MCP property schemas
for reflection/class_schema and scene/component include_schema queries.

Add std::vector<DENUM enum> codegen support in DeltaHeaderTool and wire
PropertyValueIO vector element JSON plus Details vector table enum combos.

Add MCP, PropertyValueIO, and header-tool tests for enum schema and
vector enum round-trip.


* Phase 5: adopt DENUM on DMaterial::m_renderMode (#83)

- Add DENUM() to ERenderMode and change m_renderMode to ERenderMode
- Update SetRenderMode API; remove DFUNCTION (enum params not yet supported)
- Fix tests to use enum property/reflection and ERenderMode call sites
- Add material m_renderMode integer JSON round-trip regression test
- Document DENUM reflection rules and production adoption in CLAUDE.md


* Fix reflection template

---------

+++ Engine/Tests/Engine/Graphics/Gpu/GpuTransparentRenderTests.cpp
@@ -122,7 +122,7 @@ DMaterial* GetMeshMaterial(MeshRenderer* renderer)
 void ConfigureTransparentMaterial(DMaterial* material, const XMFLOAT4& baseColor, const bool doubleSided = false)
 {
     ASSERT_NE(material, nullptr);
-    material->SetRenderMode(static_cast<uint32_t>(ERenderMode::Transparent));
+    material->SetRenderMode(ERenderMode::Transparent);
     SetFloat4Property(*material, "m_baseColor", baseColor);
     SetFloatProperty(*material, "m_metallic", 0.0f);
     SetFloatProperty(*material, "m_roughness", 0.5f);
